First React Project (Car Rental Site)
As i am still getting used to github and even newer to react i wasnt sure how to handle to project with git until today which is why there is a nearly the entire application in this first commit. However what i have built here is my first react project that i have made that i feel is worthy of my portfolio. It is a rudimental car rental site featuring a modern single page design built with the React library. To simulate getting the cars data from an API i have simply created my own seperate file from which to import the data as i could not find a good free option to use a real API for this. The rendering is done based on this data and so i could fill up the carData file with as many as i want and not have to worry about it breaking or needing new code. you can choose your car your dates and browse a selection of cars as well as multiple components about the "company" or that i felt you may find on a webpage today. The site still needs to be made responsive at this stage and there are some final functionality tweaks and info changes that will be required still to come in a later update.