Refactoring
Now that i have progressed further with react i felt it was important for me to come back to this and fix some of the glaring issue's with the first iteration.

I have updated the inputs to be controlled components and removed the vanilla JS that i had mingled in with the React code and refactored it to all be React.

I fixed a bug where the reserve button would take you to the hero section instead of the book section when clicked on a certain viewport width.